chore: move code to utils and plugins in uplot folder
