+++ frontend/src/lib/uPlotLib/uPlotLib.styles.scss
@@ -19,4 +19,5 @@
 	display: flex;
 	align-items: center;
 	gap: 8px;
+	margin-bottom: 4px;
 }
