+++ frontend/src/lib/uPlotLib/utils/getAxes.ts
@@ -1,0 +1,67 @@
+import { getToolTipValue } from 'components/Graph/yAxisConfig';
+
+import getGridColor from './getGridColor';
+
+const getAxes = (isDarkMode, yAxisUnit): any => [
+	{
+		// label: 'Date',
+		stroke: isDarkMode ? 'white' : 'black', // Color of the axis line
+		grid: {
+			stroke: getGridColor(isDarkMode), // Color of the grid lines
+			dash: [10, 10], // Dash pattern for grid lines,
+			width: 0.5, // Width of the grid lines,
+			show: true,
+		},
+		ticks: {
+			// stroke: isDarkMode ? 'white' : 'black', // Color of the tick lines
+			width: 0.3, // Width of the tick lines,
+			show: true,
+		},
+		gap: 5,
+	},
+	{
+		// label: 'Value',
+		stroke: isDarkMode ? 'white' : 'black', // Color of the axis line
+		grid: {
+			stroke: getGridColor(isDarkMode), // Color of the grid lines
+			dash: [10, 10], // Dash pattern for grid lines,
+			width: 0.3, // Width of the grid lines
+		},
+		ticks: {
+			// stroke: isDarkMode ? 'white' : 'black', // Color of the tick lines
+			width: 0.3, // Width of the tick lines
+			show: true,
+		},
+		values: (_, t): string[] =>
+			t.map((v) => {
+				const value = getToolTipValue(v.toString(), yAxisUnit);
+
+				return `${value}`;
+			}),
+		gap: 5,
+		size: (self, values, axisIdx, cycleNum): number => {
+			const axis = self.axes[axisIdx];
+
+			// bail out, force convergence
+			if (cycleNum > 1) return axis._size;
+
+			let axisSize = axis.ticks.size + axis.gap;
+
+			// find longest value
+			const longestVal = (values ?? []).reduce(
+				(acc, val) => (val.length > acc.length ? val : acc),
+				'',
+			);
+
+			if (longestVal !== '' && self) {
+				// eslint-disable-next-line prefer-destructuring, no-param-reassign
+				self.ctx.font = axis.font[0];
+				axisSize += self.ctx.measureText(longestVal).width / devicePixelRatio;
+			}
+
+			return Math.ceil(axisSize);
+		},
+	},
+];
+
+export default getAxes;
